requirements.txt
absl-py==2.1.0
asgiref==3.8.1
astroid==3.2.4
astunparse==1.6.3
blinker==1.8.2
boto3==1.34.140
botocore==1.34.140
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
coloredlogs==15.0.1
cryptography==42.0.8
Cython==3.0.10
dill==0.3.8
Django==5.0.7
django-cors-headers==4.4.0
django-extensions==3.2.3
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
dnspython==2.6.1
email_validator==2.2.0
filelock==3.15.4
Flask==3.0.3
Flask-Cors==4.0.1
Flask-JWT-Extended==4.6.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
flatbuffers==24.3.25
fsspec==2024.6.1
gast==0.6.0
gluonnlp==0.10.0
google-pasta==0.2.0
graphviz==0.8.4
grpcio==1.64.1
h5py==3.11.0
huggingface-hub==0.23.4
humanfriendly==10.0
idna==3.7
isort==5.13.2
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
keras==3.4.1
kobert-tokenizer @ git+https://github.com/SKTBrain/KoBERT.git@47a69af87928fc24e20f571fe10c3cc9dd9af9a3#subdirectory=kobert_hf
kobert-transformers==0.5.1
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mccabe==0.7.0
mdurl==0.1.2
ml-dtypes==0.3.2
mpmath==1.3.0
mxnet==1.6.0
mypy-extensions==1.0.0
mysqlclient==2.2.4
namex==0.0.8
networkx==3.3
numpy==1.26.4
onnxruntime==1.18.1
opt-einsum==3.3.0
optional-django==0.1.0
optree==0.12.0
packaging==24.1
pandas==2.2.2
pathspec==0.12.1
platformdirs==4.2.2
protobuf==4.25.3
pycodestyle==2.12.0
pycparser==2.22
pyflakes==3.2.0
Pygments==2.18.0
PyJWT==2.8.0
pylint==3.2.5
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.3
rich==13.7.1
s3transfer==0.10.2
safetensors==0.4.3
scikit-learn==1.5.1
scipy==1.14.0
sentencepiece==0.2.0
setuptools==70.2.0
six==1.16.0
SQLAlchemy==2.0.31
sqlparse==0.5.1
sympy==1.12.1
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.2
termcolor==2.4.0
tf_keras==2.16.0
threadpoolctl==3.5.0
tokenizers==0.19.1
tomlkit==0.13.0
torch==2.3.1
torchtext==0.18.0
tqdm==4.66.4
transformers==4.42.3
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
Werkzeug==3.0.3
wheel==0.43.0
wrapt==1.16.0
WTForms==3.1.2
